Moved corridor walls creation to corridorbeatification
